+++ vector_io/vector_space.go
@@ -22,11 +22,11 @@ type VectorSearcher struct{}
 func (store *vectorStore) writeVector(v *Vector) {
 	v.Id = vectorStoreInstance.LatestId + 1
 	vectorStoreInstance.LatestId++
-	store.vectorSpace.AddPoint(v.Components, v.Id)
+	store.vectorSpace.AddPoint(v.Components, uint32(v.Id))
 	store.StoredVectors[v.Id] = v
 }
 
-func (searcher *VectorSearcher) SimilaritySearch(queryVectorId t.VectorId, k uint32) ([]t.VectorId, error) {
+func (searcher *VectorSearcher) SimilaritySearch(queryVectorId t.VectorId, k int) ([]t.VectorId, error) {
 	// we perform similarity search using the HNSW algorithm with a time complexity of O(log n)
 	// when performing the algorithm, we use k+1 as the resulting k-nearest neighbors will always include the query vector itself.
 	// therefore we simply perform the search for k+1 nearest neighbors and remove the queryVectorId from the output
@@ -35,28 +35,28 @@ func (searcher *VectorSearcher) SimilaritySearch(queryVectorId t.VectorId, k uin
 	if err != nil {
 		return nil, err
 	}
-	ids, _ := vectorStoreInstance.vectorSpace.SearchKNN(queryVector.Components, int(k)+1) // returns ids of resulting vectors and the vectors' distances from the query vector
+	ids, _ := vectorStoreInstance.vectorSpace.SearchKNN(queryVector.Components, k+1) // returns ids of resulting vectors and the vectors' distances from the query vector
 
 	idsExcludingQuery := make([]t.VectorId, 0)
 	for _, id := range ids {
-		if id != queryVectorId {
-			idsExcludingQuery = append(idsExcludingQuery, id)
+		if int(id) != queryVectorId {
+			idsExcludingQuery = append(idsExcludingQuery, int(id))
 		}
 	}
 	return idsExcludingQuery, nil
 }
 
-func instantiateVectorStore(dim uint32, similarityMetric t.SimilarityMetric, spaceSize uint32, M uint32, efConstruction uint32) {
+func instantiateVectorStore(dim int, similarityMetric t.SimilarityMetric, spaceSize int, M int, efConstruction int) {
 	vectorStoreInstance = &vectorStore{}
 	vectorStoreInstance.vectorSpace = hnswgo.New(
 		int(dim),
 		int(M),
 		int(efConstruction),
 		int(time.Now().Unix()),
-		spaceSize,
+		uint32(spaceSize),
 		similarityMetric,
 	)
-	vectorStoreInstance.StoredVectors = make(map[uint32]*Vector)
+	vectorStoreInstance.StoredVectors = make(map[int]*Vector)
 
 	err := vectorStoreInstance.LoadPersistedVectors()
 	if err != nil {
